refactor: 상품 삭제시 유저에게 한번 더 확인 받도록 수정

+++ src/components/mainPage/productCard/AddCartButton.tsx
@@ -45,6 +45,11 @@ export const AddCartButton = ({ productId }: AddCartButtonProps) => {
   const deleteCartItem = () => {
     if (cartId === undefined) return;
 
+    // eslint-disable-next-line no-restricted-globals
+    const isUserWantDelete = confirm('상품을 장바구니에서 삭제하시겠습니까?');
+
+    if (!isUserWantDelete) return;
+
     deleteRecoilCartById(cartId);
     deleteCartItemById(cartId);
   };
